Unified PersonEditDialog
