+++ src/main/java/address/controller/PersonEditDialog.java
@@ -11,6 +11,7 @@ import address.util.LoggerManager;
 import javafx.application.Platform;
 import javafx.fxml.FXML;
 import javafx.fxml.FXMLLoader;
+import javafx.scene.Node;
 import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.control.Alert.AlertType;
@@ -25,6 +26,7 @@ import javafx.stage.StageStyle;
 import java.io.IOException;
 import java.net.MalformedURLException;
 import java.net.URL;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -33,14 +35,19 @@ import java.util.Optional;
  *
  * Stage, initial person and available & assigned tags should be set before showing stage
  */
-public class PersonEditDialogController extends UiController {
-
-    private static final AppLogger logger = LoggerManager.getLogger(PersonEditDialogController.class);
-    private static final String FXML_TAG_SELECTION_EDIT_DIALOG = "/view/TagSelectionEditDialog.fxml";
+public class PersonEditDialog extends BaseUiController {
+    private static final AppLogger logger = LoggerManager.getLogger(PersonEditDialog.class);
+    private static final String ICON = "/images/edit.png";
+    public static final String TITLE = "Edit Person";
+    public static final String FXML = "PersonEditDialog.fxml";
     private static final String TOOLTIP_TAG_SELECTOR_SHORTCUT = "Shortcut + O";
     private static final String TOOLTIP_LAUNCH_TAG_SELECTOR = "Click to launch tag selector";
-    private PersonEditDialogView view;
+    AnchorPane pane;
+    Stage dialogStage;
     private boolean isOkClicked = false;
+
+    //TODO: moved to a separate controller
+    private static final String FXML_TAG_SELECTION_EDIT_DIALOG = "/view/TagSelectionEditDialog.fxml";
 
     @FXML
     private Label idLabel;
@@ -77,6 +84,56 @@ public class PersonEditDialogController extends UiController {
         addListeners();
         modifyInteractionsWithComponents();
         Platform.runLater(() -> firstNameField.requestFocus());
+    }
+
+
+    @Override
+    public void setNode(Node node) {
+        pane = (AnchorPane)node;
+    }
+
+    @Override
+    public String getFxmlPath() {
+        return FXML;
+    }
+
+    @Override
+    public void secondaryInit() {
+        Scene scene = new Scene(pane);
+        dialogStage = loadDialogStage(TITLE, primaryStage, scene);
+        setIcon(dialogStage, ICON);
+        setEscKeyToDismiss(scene);
+    }
+
+    private void setEscKeyToDismiss(Scene scene) { //TODO: move to a new parent class BaseDialogView
+        scene.setOnKeyPressed(event -> {
+            if (event.getCode() == KeyCode.ESCAPE) {
+                dialogStage.close();
+            }
+        });
+    }
+
+    public void showAndWait(){
+        dialogStage.showAndWait();
+    }
+
+    public void setData(ReadOnlyPerson initialData, List<Tag> tags) {
+        setInitialPersonData(initialData);
+        setTags(tags, new ArrayList<>(initialData.getObservableTagList()));
+    }
+
+    public Optional<ReadOnlyPerson> getUserInput() {
+        showAndWait();
+        if (isOkClicked()) {
+            logger.debug("Person collected: " + getEditedPerson().toString());
+            return Optional.of(getEditedPerson());
+        } else {
+            return Optional.empty();
+        }
+    }
+
+    public void close() {
+        dialogStage.close();
     }
 
     private void modifyInteractionsWithComponents() {
@@ -122,7 +179,7 @@ public class PersonEditDialogController extends UiController {
             // Create the dialog Stage.
             Stage dialogStage = new Stage();
             dialogStage.initModality(Modality.WINDOW_MODAL);
-            dialogStage.initOwner(view.getDialogStage());
+            dialogStage.initOwner(dialogStage);
             dialogStage.initStyle(StageStyle.TRANSPARENT);
 
             Scene scene = new Scene(pane, Color.TRANSPARENT);
@@ -140,7 +197,7 @@ public class PersonEditDialogController extends UiController {
             logger.warn("Error launching tag selection dialog: {}", e);
 
             Alert alert = new Alert(AlertType.ERROR);
-            alert.initOwner(view.getDialogStage());
+            alert.initOwner(dialogStage);
             alert.setTitle("FXML Load Error");
             alert.setHeaderText("Cannot load dialog for tag selection dialog");
             alert.setContentText("IOException when trying to load " + FXML_TAG_SELECTION_EDIT_DIALOG);
@@ -206,13 +263,13 @@ public class PersonEditDialogController extends UiController {
         finalPerson.setTags(finalAssignedTags);
         finalPerson.setGithubUsername(githubUserNameField.getText());
         isOkClicked = true;
-        view.close();
+        close();
     }
 
     private void showInvalidInputMessage(String errorMessage) {
         // Show the error message.
         Alert alert = new Alert(AlertType.ERROR);
-        alert.initOwner(view.getDialogStage());
+        alert.initOwner(dialogStage);
         alert.setTitle("Invalid Fields");
         alert.setHeaderText("Please correct invalid fields");
         alert.setContentText(errorMessage);
@@ -230,7 +287,7 @@ public class PersonEditDialogController extends UiController {
      */
     @FXML
     protected void handleCancel() {
-        view.close();
+        close();
     }
 
     /**
@@ -272,11 +329,8 @@ public class PersonEditDialogController extends UiController {
         return textField.getText() != null && textField.getText().length() != 0;
     }
 
-    public void setView(PersonEditDialogView view) {
-        this.view = view;
-    }
-
     public boolean isOkClicked() {
         return isOkClicked;
     }
+
 }
