+++ src/main/resources/view/PersonEditDialog.fxml
@@ -8,7 +8,7 @@
 <?import java.net.URL?>
 
 <AnchorPane fx:id="mainPane" prefHeight="404.0" prefWidth="439.0" styleClass="background" xmlns="http://javafx.com/javafx/8"
-            xmlns:fx="http://javafx.com/fxml/1" fx:controller="address.controller.PersonEditDialogController">
+            xmlns:fx="http://javafx.com/fxml/1" fx:controller="address.controller.PersonEditDialog">
     <stylesheets>
         <URL value="@DarkTheme.css" />
         <URL value="@Extensions.css" />
